+++ hydra/blocks/spinhalf/terms/apply_exchange.h
@@ -23,8 +23,8 @@ void apply_exchange(Bond const &bond, IndexingIn &&indexing_in,
   assert(bond.sites_disjoint());
 
   coeff_t J = bond.coupling<coeff_t>();
-  int s1 = bond[0];
-  int s2 = bond[1];
+  int64_t s1 = bond[0];
+  int64_t s2 = bond[1];
   bit_t flipmask = ((bit_t)1 << s1) | ((bit_t)1 << s2);
 
   // Define actions of bond
