+++ hydra/blocks/spinhalf/terms/apply_scalar_chirality.h
@@ -34,9 +34,9 @@ void apply_scalar_chirality(Bond const &bond, IndexingIn &&indexing_in,
     Jquarter_conj = hydra::conj(Jquarter);
   }
 
-  int s1 = bond[0];
-  int s2 = bond[1];
-  int s3 = bond[2];
+  int64_t s1 = bond[0];
+  int64_t s2 = bond[1];
+  int64_t s3 = bond[2];
   bit_t spinmask = ((bit_t)1 << s1) | ((bit_t)1 << s2) | ((bit_t)1 << s3);
 
   // scalar chirality annihilates 000 and 111
