+++ hydra/blocks/tj/tj_utils.h
@@ -5,7 +5,7 @@
 namespace hydra::utils {
 
 template <typename bit_t>
-constexpr int tj_site_compressed(bit_t holes, int site) {
+constexpr int64_t tj_site_compressed(bit_t holes, int64_t site) {
   return site - bitops::popcnt(holes & (((bit_t)1 << site) - 1));
 }
 
@@ -30,7 +30,7 @@ constexpr std::pair<bit_t, bit_t> up_dnc_to_upc_dn(bit_t up, bit_t dnc) {
 
 template <typename bit_t>
 constexpr std::pair<bit_t, bit_t> holes_spins_to_up_dn(bit_t holes, bit_t spins,
-                                                       int n_sites) {
+                                                       int64_t n_sites) {
   bit_t mask = ((bit_t)1 << n_sites) - 1;
   bit_t up = bitops::deposit(spins, (bit_t)~holes);
   bit_t dn = bitops::deposit((bit_t)~spins, (bit_t)~holes) & mask;
@@ -39,7 +39,7 @@ constexpr std::pair<bit_t, bit_t> holes_spins_to_up_dn(bit_t holes, bit_t spins,
 
 template <typename bit_t>
 constexpr std::pair<bit_t, bit_t> up_dn_to_holes_spins(bit_t up, bit_t dn,
-                                                       int n_sites) {
+                                                       int64_t n_sites) {
   bit_t mask = ((bit_t)1 << n_sites) - 1;
   bit_t holes = (~(up | dn)) & mask;
   bit_t spins = bitops::extract(up, (bit_t)~holes);
