+++ hydra/operators/non_branching_bonds.h
@@ -19,10 +19,10 @@ public:
   bool non_zero_term(bit_t local_state) const;
   bit_t extract_local_state(bit_t state) const;
   bit_t deposit_local_state(bit_t local_state, bit_t state) const;
-  int number_difference() const;
+  int64_t number_difference() const;
 
 private:
-  std::vector<int> sites_;
+  std::vector<int64_t> sites_;
   bit_t dim_;
   bit_t mask_;
   arma::cx_mat matrix_;
